requirements.txt
Django==5.2
python-telegram-bot==13.15
environs==14.5.0
urllib3==1.26.18
